Add minimum tox + unit tests for wrapper

+++ src/benchmark/wrapper/process.py
@@ -27,6 +27,9 @@ VALID_LOG_LEVELS = ["info", "debug", "warning", "error", "critical"]
 
 
 logger = logging.getLogger(__name__)
+logging.basicConfig(
+    filename="/var/log/dpe_benchmark_workload.log", encoding="utf-8", level=logging.INFO
+)
 
 
 class BenchmarkProcess:
@@ -42,7 +45,7 @@ class BenchmarkProcess:
 
     def __init__(
         self,
-        model: ProcessModel,
+        model: ProcessModel | None,
         args: WorkloadCLIArgsModel,
         metrics: BenchmarkMetrics,
     ):
@@ -53,6 +56,8 @@ class BenchmarkProcess:
 
     def start(self):
         """Start the process."""
+        if not self.model:
+            return
         self._proc = subprocess.Popen(
             self.model.cmd,
             user=self.model.user,
@@ -73,6 +78,10 @@ class BenchmarkProcess:
     def status(self) -> ProcessStatus:
         """Return the status of the process."""
         stat = ProcessStatus.STOPPED
+        if not self._proc:
+            # We are managing only, we do not run a process
+            return ProcessStatus.RUNNING
+
         if self._proc.poll() is None:
             stat = ProcessStatus.RUNNING
         elif self._proc.returncode != 0:
@@ -95,20 +104,21 @@ class BenchmarkProcess:
             or (self.status() == ProcessStatus.RUNNING and self.args.duration == 0)
         ):
             to_wait = True
-            for line in iter(self._proc.stdout.readline, ""):
-                if output := self.process_line(line):
-                    self.metrics.add(output)
+            if self._proc:
+                for line in iter(self._proc.stdout.readline, ""):
+                    if output := self.process_line(line):
+                        self.metrics.add(output)
 
-                    if self.status() != ProcessStatus.RUNNING:
-                        # Process has finished
-                        break
+                        if self.status() != ProcessStatus.RUNNING:
+                            # Process has finished
+                            break
 
-                    to_wait = False
+                        to_wait = False
 
-                # Log the output.
-                # This way, an user can see what the process is doing and
-                # some of the metrics will be readily available without COS.
-                logger.info(f"[workload pid {self._proc.pid}] " + line.rstrip())
+                    # Log the output.
+                    # This way, an user can see what the process is doing and
+                    # some of the metrics will be readily available without COS.
+                    logger.info(f"[workload pid {self._proc.pid}] " + line.rstrip())
 
             if to_wait:
                 # In case the stdout is empty, we ensure we sleep anyways
@@ -126,7 +136,8 @@ class BenchmarkProcess:
     def stop(self):
         """Stop the process."""
         try:
-            self._proc.kill()
+            if self._proc:
+                self._proc.kill()
         except Exception as e:
             logger.warning(f"Error stopping worker: {e}")
         self.model.status = ProcessStatus.STOPPED
@@ -214,16 +225,16 @@ class WorkloadToProcessMapping(ABC):
         raise ValueError(f"Invalid command: {cmd}")
 
     @abstractmethod
-    def _map_prepare(self) -> tuple[BenchmarkManager, list[BenchmarkProcess] | None]:
+    def _map_prepare(self) -> tuple[BenchmarkManager | None, list[BenchmarkProcess] | None]:
         """Returns the mapping for the prepare phase."""
         ...
 
     @abstractmethod
-    def _map_run(self) -> tuple[BenchmarkManager, list[BenchmarkProcess] | None]:
+    def _map_run(self) -> tuple[BenchmarkManager | None, list[BenchmarkProcess] | None]:
         """Returns the mapping for the run phase."""
         ...
 
     @abstractmethod
-    def _map_clean(self) -> tuple[BenchmarkManager, list[BenchmarkProcess] | None]:
+    def _map_clean(self) -> tuple[BenchmarkManager | None, list[BenchmarkProcess] | None]:
         """Returns the mapping for the clean phase."""
         ...
